auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Auth service: registration + JWT issuing"
readme = "README.md"
authors = [
    { name = "natpavge", email = "natpav.ge@gmail.com" }
]
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "ruff>=0.14.0",
    "greenlet>=3.3.1",
    "bcrypt==4.3.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "pytest-cov>=7.1.0",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"


[tool.uv]
package = false

[tool.ruff]
line-length = 320
target-version = "py312"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
